+++ src/core/mod.rs
@@ -2,3 +2,4 @@ pub mod fmt;
 pub mod parse;
 pub mod scan;
 pub mod spec;
+pub mod data;
