+++ src/core/scan/cdfa.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 use std::collections::HashSet;
-use std::collections::LinkedList;
-use std::collections::linked_list::Iter;
+use core::data::stream::ReadDrivenStream;
 use std::iter::Peekable;
 use std::error;
 use std::fmt;
@@ -178,64 +177,5 @@ impl<V: Default> CEHashMap<V> {
 
     fn contains(&self, key: usize) -> bool {
         self.get(key).is_some()
-    }
-}
-
-
-//TODO separate into its own data type file with tests
-pub struct ReadDrivenStream<T: Clone> {
-    incoming_buffer: LinkedList<T>,
-    outgoing_buffer: LinkedList<T>,
-    getter: fn() -> Option<T>,
-    block: bool,
-}
-
-impl<T: Clone> ReadDrivenStream<T> {
-    fn observe(getter: fn() -> Option<T>) -> ReadDrivenStream<T> {
-        ReadDrivenStream {
-            incoming_buffer: LinkedList::new(),
-            outgoing_buffer: LinkedList::new(),
-            getter,
-            block: false,
-        }
-    }
-
-    fn pull(&mut self) -> Option<T> {
-        let val: T = match self.incoming_buffer.pop_back() {
-            None => match (self.getter)() {
-                None => { return None; }
-                Some(val) => val
-            },
-            Some(val) => val
-        };
-
-        self.outgoing_buffer.push_back(val.clone());
-
-        Some(val)
-    }
-
-    fn replay(&mut self) {
-        let mut val: Option<T> = self.outgoing_buffer.pop_front();
-        while val.is_some() {
-            self.incoming_buffer.push_back(val.unwrap());
-            val = self.outgoing_buffer.pop_front();
-        }
-    }
-
-    fn consume(&mut self) {
-        if self.block {
-            self.replay();
-            self.block = false;
-        } else {
-            self.outgoing_buffer.clear();
-        }
-    }
-
-    fn block(&mut self) {
-        self.block = true;
-    }
-
-    fn unblock(&mut self) {
-        self.block = false;
     }
 }
